Move graph and path finding to global namespace
