+++ src/navigation/global/global_planner.hh
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 #include "eigen3/Eigen/Dense"
-#include "navigation/map_graph.hh"
+#include "navigation/global/map_graph.hh"
 
 namespace navigation {
 namespace global {
