+++ src/navigation/global/path_finding.hh
@@ -1,12 +1,13 @@
-#ifndef SRC_NAVIGATION_PATH_FINDING_HH_
-#define SRC_NAVIGATION_PATH_FINDING_HH_
+#ifndef SRC_NAVIGATION_GLOBAL_PATH_FINDING_HH_
+#define SRC_NAVIGATION_GLOBAL_PATH_FINDING_HH_
 
-#include "map_graph.hh"
+#include "navigation/global/map_graph.hh"
 
 #include <vector>
 #include "eigen3/Eigen/Dense"
 
 namespace navigation {
+namespace global {
 
 std::vector<Eigen::Vector2f> dijkstra(MapGraph& graph,
                                       const Eigen::Vector2f& start,
@@ -16,6 +17,7 @@ std::vector<Eigen::Vector2f> astar(MapGraph& graph,
                                    const Eigen::Vector2f& start,
                                    const Eigen::Vector2f& goal);
 
+}  // namespace global
 }  // namespace navigation
 
-#endif  // SRC_NAVIGATION_PATH_FINDING_HH_
+#endif  // SRC_NAVIGATION_GLOBAL_PATH_FINDING_HH_
